backend/requirements.txt
flask
flask-cors
chromadb
pypdf
python-dotenv
requests
sentence-transformers